services/memory/requirements.txt
fastapi==0.103.0
pydantic==2.4.2
redis==4.5.4
python-dotenv==1.0.0
prometheus-client==0.17.1
uvicorn==0.23.2